Suppress compiler warnings in charset-detector
charset-detector generates quite a few compiler warnings when building
on Ubuntu Intrepid. This commit changes the code to suppress those
warnings.

svn path=/trunk/extragear/multimedia/amarok/; revision=934773

+++ src/charset-detector/src/nsSBCharSetProber.cpp
@@ -59,9 +59,9 @@ nsProbingState nsSingleByteCharSetProber::HandleData(const char* aBuf, PRUint32 
       {
         mTotalSeqs++;
         if (!mReversed)
-          ++(mSeqCounters[mModel->precedenceMatrix[mLastOrder*SAMPLE_SIZE+order]]);
+          ++(mSeqCounters[int(mModel->precedenceMatrix[mLastOrder*SAMPLE_SIZE+order])]);
         else // reverse the order of the letters in the lookup
-          ++(mSeqCounters[mModel->precedenceMatrix[order*SAMPLE_SIZE+mLastOrder]]);
+          ++(mSeqCounters[int(mModel->precedenceMatrix[order*SAMPLE_SIZE+mLastOrder])]);
       }
     }
     mLastOrder = order;
